Add files via upload
Queries to use in ORM.
pullMajors - Used to pull a list of majors based on a specified institute.
deleteMaps - Used to delete from both MAPS_DMAPS_META, and MAPS_DMAPS_LISTS given a specified map ID,
insertMapBlank - Used to insert into both MAPS_DMAPS_META and MAPS_DMAPS_LISTS given various inputs.
insertMapTest - Used as a test script for a filled in insertMapBlank.
findCourse - Used on the admin side - based on institute_id, the selected major, and an input string, returns all courses that match the Major/Institute and input. Input can either be a catalog number or a substring of the course description.